+++ loraham_daemon/radio_controller.h
@@ -9,6 +9,7 @@
 
 #include "radio_channel.h"
 #include "daemon_stats.h"
+#include "daemon_tx_worker.h"
 #include "radio_health.h"
 
 /* --- Radio hardware/runtime state --------------------------------------- */
@@ -47,6 +48,7 @@ struct RadioController {
     std::atomic<bool> tx_result_active;
     RadioTxMode_t tx_mode;
     uint16_t tx_result_seq;
+    DaemonTxWorker tx_worker;
 
     DaemonRadioStats stats;
 
@@ -83,6 +85,7 @@ static inline void radio_controller_init(RadioController<RadioT> *ctrl,
     ctrl->tx_result_active.store(false);
     ctrl->tx_mode = RADIO_TX_MODE_MANAGED;
     ctrl->tx_result_seq = 0;
+    daemon_tx_worker_init(&ctrl->tx_worker);
 
     daemon_radio_stats_init(&ctrl->stats);